single server: postgres exporter uses common datasource fn
