+++ internal/db/dbutil/dbutil.go
@@ -292,3 +292,52 @@ func (n JSONInt64Set) Value() (driver.Value, error) {
 	}
 	return *n.Set, nil
 }
+
+func PostgresDSN(currentUser string, getenv func(string) string) string {
+	// PGDATASOURCE is a sourcegraph specific variable for just setting the DSN
+	if dsn := getenv("PGDATASOURCE"); dsn != "" {
+		return dsn
+	}
+
+	// TODO match logic in lib/pq
+	// https://sourcegraph.com/github.com/lib/pq@d6156e141ac6c06345c7c73f450987a9ed4b751f/-/blob/connector.go#L42
+	dsn := &url.URL{
+		Scheme: "postgres",
+		Host:   "127.0.0.1:5432",
+	}
+
+	// Username preference: PGUSER, $USER, postgres
+	username := "postgres"
+	if currentUser != "" {
+		username = currentUser
+	}
+	if user := getenv("PGUSER"); user != "" {
+		username = user
+	}
+
+	if password := getenv("PGPASSWORD"); password != "" {
+		dsn.User = url.UserPassword(username, password)
+	} else {
+		dsn.User = url.User(username)
+	}
+
+	if host := getenv("PGHOST"); host != "" {
+		dsn.Host = host
+	}
+
+	if port := getenv("PGPORT"); port != "" {
+		dsn.Host += ":" + port
+	}
+
+	if db := getenv("PGDATABASE"); db != "" {
+		dsn.Path = db
+	}
+
+	if sslmode := getenv("PGSSLMODE"); sslmode != "" {
+		qry := dsn.Query()
+		qry.Set("sslmode", sslmode)
+		dsn.RawQuery = qry.Encode()
+	}
+
+	return dsn.String()
+}
